search: move searcher-only logic to searcher package
